Modify to only support HCL write in Go 1.7
